% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_score.R
\name{apache_score}
\alias{apache_score}
\title{Calculate APACHE-II score}
\usage{
apache_score(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A data frame with all original columns plus APACHE II score
}
\description{
\code{apache_score} takes a data frame with patient data and calculates the
APACHE II score
}
\details{
This function takes a data frame containing all of the data needed to
calculate the APAPCHE II score.
}

