% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{calc_runtime}
\alias{calc_runtime}
\title{Calculate the running time for continuous medication data}
\usage{
calc_runtime(cont.data, units = "hours")
}
\arguments{
\item{cont.data}{A data frame with continuous medication rate data}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
A data frame
}
\description{
\code{calc_runtime} calculates the time at current rate and time from
start
}
\details{
This function takes a data frame with continuous medication rate data and
produces a data frame with the time at each rate and the time from start for
each row. This could be used to then calculate the AUC or to summarize the
continuous data.
}

