% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{calc_unit_los}
\alias{calc_unit_los}
\title{Calculate the length of stay in each unit}
\usage{
calc_unit_los(unit.data, units = "days")
}
\arguments{
\item{unit.data}{A data frame with hospital unit arrival and departure
information}

\item{units}{An optional character string specifying the time units to use in
calculations, default is days}
}
\value{
A data frame
}
\description{
\code{calc_unit_los} calculates the length of stay in each hospital unit
}
\details{
This function takes a data frame with arrival and departure times and
determines the length of stay in each unit. This accounts for data which
might contain duplicate arrival or departure information.
}

