% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rowsback.R
\name{count_rowsback}
\alias{count_rowsback}
\title{Count the number of rows to go back in data frame}
\usage{
count_rowsback(datetime, back = 2)
}
\arguments{
\item{datetime}{A vector of POSIXct values}

\item{back}{An optional numeric specifying the number of days back to go.
Defaults to 2 days.}
}
\value{
An integer with the  number of rows; can be passed on to
  \code{\link[zoo]{rollapplyr}}
}
\description{
\code{count_rowsback} determines how many rows should be included
}
\details{
This function takes a vector of date/times (of type POSIXct) and counts the
number of rows which would fall within the specified time frame. Typically
called from \code{\link[dplyr]{mutate}} and the results are passed on to
\code{\link[zoo]{rollapplyr}}.
}

