% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{fill_false}
\alias{fill_false}
\title{Fill NA with FALSE}
\usage{
fill_false(y)
}
\arguments{
\item{y}{A value}
}
\value{
Either FALSE or the original value
}
\description{
\code{fill_false} returns FALSE if the value is NA, otherwise returns the
original value
}
\details{
This function takes a value and checks if it is NA. If it is NA it will
return FALSE, otherwise it will return the original value. This can be used
to fill data frame rows which are all NA but should be FALSE. with reference
diagnosis codes and a data
}

