% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_dates}
\alias{filter_dates}
\title{Filter values to specified date range}
\usage{
filter_dates(data, criteria, prior.start = 0, after.end = 0,
  dtcols = c("datetime", "start.datetime", "stop.datetime"))
}
\arguments{
\item{data}{A data frame with a date/time column called datetime}

\item{criteria}{A data frame with start and stop date/time columns called
start.datetime and end.datetime}

\item{prior.start}{An optional numeric indicating number of days to go back
prior to the start date/time, default is 0}

\item{after.end}{An optional numeric indicating number of days after the end
date/time to include, default is 0}

\item{dtcols}{An optional vector of strings with the names of the columns to
be used for filtering}
}
\value{
A data frame
}
\description{
\code{filter_dates} filters data to within a specified time frame
}
\details{
This function takes a data frame with a column called \code{datetime} and a
data frame with columns called \code{start.datetime} and \code{end.datetime},
and filters the data to only include values within the specified time frame.
Optionally, two additional parameters (prior.start and after.end) can be
included to add additional time before or after the start or end date and
time. Both data frames should include a pie.id column to use for joining.
}

