% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9_lookup.R
\name{icd9_description}
\alias{icd9_description}
\title{Lookup ICD9 code description}
\usage{
icd9_description(codes)
}
\arguments{
\item{codes}{A character vector of ICD9 codes}
}
\value{
A data frame with columns: icd9.code and icd9.description
}
\description{
\code{icd9_description} takes a vector of ICD9 codes and returns a data frame
with the corresponding description for each code
}
\details{
This function takes a character vector of ICD9 codes and returns a data frame
with two columns: icd9.code and icd9.description.
}

