% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9_lookup.R
\name{icd9_lookup}
\alias{icd9_lookup}
\title{Lookup ICD9 codes from CCS codes}
\usage{
icd9_lookup(df)
}
\arguments{
\item{df}{A data frame with columns: disease.state, type, code}
}
\value{
A data frame with columns: disease.state, icd9.code
}
\description{
\code{icd9_lookup} takes a data frame with CCS codes and returns the
corresponding ICD9 codes
}
\details{
This function takes a data frame with three columns: disease.state, type, and
code. The column \code{disease.state} is a character field with the name of a
disease state which will be used for grouping. The column \code{type} is a
character field with either "ICD9" or "CCS", which indicates the type of
code. The column \code{code} is a character field with the ICD9 or CCS code.
For all rows with CCS codes, the function will look-up the corresponding ICD9
code and then return a data frame with two columns: disease.state and
icd9.code.
}

