% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{lab_change}
\alias{lab_change}
\title{Determine if a lab changed by a set amount within a specific time frame}
\usage{
lab_change(lab.data, back = 2, units = "days")
}
\arguments{
\item{lab.data}{A data frame with continuous medication rate data}

\item{back}{An optional numeric specifying the number of days back to go.
Defaults to 2 days.}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
A data frame
}
\description{
\code{lab_change} checks for changes in lab values
}
\details{
This function takes a data frame with continuous medication rate data and
produces a data frame with summary data for each patient and medication. The
calculations include: first rate, last rate, minimum rate, maximum rate, AUC,
time-weighted average rate, total infusion duration, total infusion running
time, and cumulative dose.
}

