% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edw_data.R
\name{read_data}
\alias{read_data}
\title{Read and join data from multiple csv files}
\usage{
read_data(data.dir, file.name)
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing the
data files}

\item{file.name}{A character string with name of data file or pattern to match}
}
\value{
A data frame
}
\description{
\code{read_data} takes a directory and file name and reads in all matching
csv files and binds them together into a data frame
}
\details{
This function takes a directory and file name and reads in all matching csv
files and binds them together into a data frame.
}

