% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edw_data.R
\name{read_edw_data}
\alias{read_edw_data}
\title{Read EDW data from csv files}
\usage{
read_edw_data(data.dir, file.name, type = NA)
}
\arguments{
\item{data.dir}{A character with the name of the directory containing the
data files}

\item{type}{An optional character indicating type of data being tidied}

\item{pattern}{A character with name of data file or pattern to match}
}
\value{
A data frame
}
\description{
\code{read_edw_data} takes a directory and file name and reads in all
matching csv files and binds them together into a data frame
}
\details{
This function takes a directory and file name and reads in all
matching csv files and binds them together into a data frame.
}

