% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_cont_meds}
\alias{summarize_cont_meds}
\title{Summary calculations for continuous medication data}
\usage{
summarize_cont_meds(cont.data, units = "hours")
}
\arguments{
\item{cont.data}{A data frame with continuous medication rate data}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
A data frame
}
\description{
\code{summarize_cont_meds} summarizes continuous medication data
}
\details{
This function takes a data frame with continuous medication rate data and
produces a data frame with summary data for each patient and medication. The
calculations include: first rate, last rate, minimum rate, maximum rate, AUC,
time-weighted average rate, total infusion duration, total infusion running
time, and cumulative dose.
}

