% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy data}
\usage{
tidy_data(type, ...)
}
\arguments{
\item{type}{A character indicating what type of data is being tidied}

\item{...}{parameters to pass on to the underlying tidy function}
}
\value{
A data frame
}
\description{
\code{tidy_data} tidy data from standard EDW queries
}
\details{
This function calls the underlying tidy function based on the value passed to
the type parameter and returns the tidy data frame. Valid options for type
are: diagnosis, meds_cont, meds_outpt.
}

