% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_diagnosis}
\alias{tidy_diagnosis}
\title{Tidy diagnosis codes}
\usage{
tidy_diagnosis(ref.data, pt.data, patients)
}
\arguments{
\item{ref.data}{A data frame with the desired diagnosis codes}

\item{pt.data}{A data frame with all patient diagnosis codes}

\item{patients}{A data frame with a column pie.id including all patients in study}
}
\value{
A data frame
}
\description{
\code{tidy_diagnosis} determines which patients have the desired diagnosis
}
\details{
This function takes a data frame with reference diagnosis codes and a data
frame with all patient diagnosis codes, and returns a data frame with a
logical for each disease state for each patient.
}

