% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edw_data.R
\name{tidy_edw_data}
\alias{tidy_edw_data}
\title{Tidy data from EDW}
\usage{
tidy_edw_data(raw.data, type)
}
\arguments{
\item{raw.data}{A data frame with raw data from EDW}

\item{type}{A character vector indicating the type of data being tidied}
}
\value{
A data frame
}
\description{
\code{tidy_edw_data} takes a data frame with raw data from EDW and returns a
tidied data frame
}
\details{
This function takes a data frame with raw data generated by an EDW query, and
a character vector indicating the type of data being tidied, and will return
a tidied data frame. Valid options for type include: blood, demographics,
diagnosis, ht_wt, home_meds, icu_score, labs_cbc, labs_chemistry, labs_coag,
labs_hit, labs_lft, labs_preg, labs_renal, location, meds_cont, meds_sched,
mpp, procedure, surgery, radiology, vent, vitals, warfarin.
}

