% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_meds_sched}
\alias{tidy_meds_sched}
\title{Tidy scheduled medications}
\usage{
tidy_meds_sched(ref.data, sched.data, patients)
}
\arguments{
\item{ref.data}{A data frame with three columns: name, type, and group}

\item{sched.data}{A data frame with all scheduled medications}

\item{patients}{A data frame with a column pie.id including all patients in
study}
}
\value{
A data frame
}
\description{
\code{tidy_meds_sched} determines which patients have the desired medications
}
\details{
This function takes a data frame with reference medications or medication
classes and a data frames with all scheduled medications, and
returns a data frame with only the desired medications for each patient. The
data frame passed to ref.data should contain three columns: name, type, and
group. The name column should contain either generic medication names or
medication classes. The type column should specify whether the value in name
is a "class" or "med". The group column should specify whether the medication
is a continous or scheduled medication.
}

