% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{lab_change}
\alias{lab_change}
\title{Determine if a lab changed by a set amount within a specific time frame}
\usage{
lab_change(lab.data, change.by, FUN, back = 2)
}
\arguments{
\item{lab.data}{A data frame with lab data}

\item{change.by}{A numeric indicating the threshold for lab changes}

\item{FUN}{A function for rollapplyr, most commonly max or min}

\item{back}{An optional numeric specifying the number of days back to go.
Defaults to 2 days.}
}
\value{
A data frame
}
\description{
\code{lab_change} checks for changes in lab values
}
\details{
This function takes a data frame with lab data for a single lab and checks
whether the lab changes by a certain amount within a given period of time.
The parameters should include: change.by, the threshold which the lab must
change by; FUN, the function passed to rollapplyr; back, the time frame that
the lab change must occur in. For FUN, use max when looking for a decrease in
lab value, and min when looking for an increase in lab value.
}
\examples{
\dontrun{
lab_change(data, -2, max, back = 2)
# checks for a >= 2 decrease in the lab value within the past 2 days
}

}

